Implement synchronous script execution in WebView2
Ok, here's my attempt to solve the sync issue.

Unfortunately, all my attempts to add a similar solution for window.OnInitialize and CallJS failed, both create a deadlock when implemented this way.

However, this appears to solve the issue on window.OnUpdate which I think is the most important.

Please build my Working2 branch and test it, you should see now both Status: Update #N on Rainmeter and on WebView are correctly synced. As well as the Measure's string value.

Also add
OnPageLoadAction=[!UpdateMeter *][!Redraw]

on the WebView measure to correctly see the "Initialized!" message on the status string.